+++ aws/cloudtrail/getter.go
@@ -2,11 +2,11 @@ package cloudtrail
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/cloudtrail"
 	"github.com/aws/aws-sdk-go-v2/service/cloudtrail/types"
+	"github.com/padok-team/yatas-aws/logger"
 )
 
 func GetCloudtrails(s aws.Config) []types.Trail {
@@ -16,7 +16,9 @@ func GetCloudtrails(s aws.Config) []types.Trail {
 	}
 	result, err := svc.DescribeTrails(context.TODO(), input)
 	if err != nil {
-		fmt.Println(err)
+		logger.Logger.Error(err.Error())
+		// Return an empty list
+		return []types.Trail{}
 	}
 	return result.TrailList
 }