+++ aws/ec2/getter.go
@@ -2,10 +2,10 @@ package ec2
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
+	"github.com/padok-team/yatas-aws/logger"
 )
 
 type EC2GetObjectAPI interface {
@@ -16,7 +16,9 @@ func GetEC2s(svc EC2GetObjectAPI) []types.Instance {
 	input := &ec2.DescribeInstancesInput{}
 	result, err := svc.DescribeInstances(context.TODO(), input)
 	if err != nil {
-		fmt.Println(err)
+		logger.Logger.Error(err.Error())
+		// Return an empty list
+		return []types.Instance{}
 	}
 	var instances []types.Instance
 	for _, r := range result.Reservations {
@@ -29,7 +31,9 @@ func GetEC2s(svc EC2GetObjectAPI) []types.Instance {
 		input.NextToken = result.NextToken
 		result, err = svc.DescribeInstances(context.TODO(), input)
 		if err != nil {
-			fmt.Println(err)
+			logger.Logger.Error(err.Error())
+			// Return an empty list of instances
+			return []types.Instance{}
 		}
 		for _, r := range result.Reservations {
 			instances = append(instances, r.Instances...)