+++ aws/eks/getter.go
@@ -2,11 +2,11 @@ package eks
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/eks"
 	"github.com/aws/aws-sdk-go-v2/service/eks/types"
+	"github.com/padok-team/yatas-aws/logger"
 )
 
 type EKSGetObjectAPI interface {
@@ -18,7 +18,9 @@ func GetClusters(svc EKSGetObjectAPI) []types.Cluster {
 	input := &eks.ListClustersInput{}
 	result, err := svc.ListClusters(context.TODO(), input)
 	if err != nil {
-		fmt.Println(err)
+		logger.Logger.Error(err.Error())
+		// Return an empty list
+		return []types.Cluster{}
 	}
 	var clusters []string
 	var clustersDetails []types.Cluster
@@ -32,7 +34,9 @@ func GetClusters(svc EKSGetObjectAPI) []types.Cluster {
 		input.NextToken = result.NextToken
 		result, err = svc.ListClusters(context.TODO(), input)
 		if err != nil {
-			fmt.Println(err)
+			logger.Logger.Error(err.Error())
+			// Return an empty list of instances
+			return []types.Cluster{}
 		}
 		for _, r := range result.Clusters {
 			clusters = append(clusters, r)
@@ -45,7 +49,9 @@ func GetClusters(svc EKSGetObjectAPI) []types.Cluster {
 		}
 		result, err := svc.DescribeCluster(context.TODO(), input)
 		if err != nil {
-			fmt.Println(err)
+			logger.Logger.Error(err.Error())
+			// Return an empty list of instances
+			return []types.Cluster{}
 		}
 		clustersDetails = append(clustersDetails, *result.Cluster)
 	}