+++ aws/lambda/getter.go
@@ -2,11 +2,11 @@ package lambda
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
 	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
+	"github.com/padok-team/yatas-aws/logger"
 )
 
 func GetLambdas(s aws.Config) []types.FunctionConfiguration {
@@ -16,18 +16,22 @@ func GetLambdas(s aws.Config) []types.FunctionConfiguration {
 		MaxItems: aws.Int32(100),
 	}
 	result, err := svc.ListFunctions(context.TODO(), input)
-	lambdas = append(lambdas, result.Functions...)
 	if err != nil {
-		fmt.Println(err)
+		logger.Logger.Error(err.Error())
+		// Return an empty list
+		return []types.FunctionConfiguration{}
 	}
+	lambdas = append(lambdas, result.Functions...)
 	for {
 		if result.NextMarker != nil {
 			input.Marker = result.NextMarker
 			result, err = svc.ListFunctions(context.TODO(), input)
-			lambdas = append(lambdas, result.Functions...)
 			if err != nil {
-				fmt.Println(err)
+				logger.Logger.Error(err.Error())
+				// Return an empty list
+				return []types.FunctionConfiguration{}
 			}
+			lambdas = append(lambdas, result.Functions...)
 		} else {
 			break
 		}