+++ aws/s3/s3.go
@@ -10,7 +10,8 @@ import (
 	"github.com/padok-team/yatas/plugins/commons"
 )
 
-func CheckS3Location(s aws.Config, bucket, region string) bool {
+// Return true if the bucket is in the region, false with the correct region if not
+func CheckS3Location(s aws.Config, bucket, region string) (bool, string) {
 
 	svc := s3.NewFromConfig(s)
 
@@ -19,19 +20,19 @@ func CheckS3Location(s aws.Config, bucket, region string) bool {
 	}
 	resp, err := svc.GetBucketLocation(context.TODO(), params)
 	if err != nil {
-
-		return false
+		return false, ""
 	}
 
-	if resp.LocationConstraint != "" {
-		if string(resp.LocationConstraint) == region {
-			return true
+	if string(resp.LocationConstraint) == region {
+		return true, region
+	} else if string(resp.LocationConstraint) == "" { // If the bucket is in us-east-1, the LocationConstraint is empty
+		if region == "us-east-1" {
+			return true, region
 		} else {
-			return false
+			return false, "us-east-1"
 		}
-
 	} else {
-		return false
+		return false, string(resp.LocationConstraint)
 	}
 }
 
@@ -43,7 +44,7 @@ type BucketAndNotInRegion struct {
 func RunChecks(wa *sync.WaitGroup, s aws.Config, c *commons.Config, queue chan []commons.Check) {
 
 	var checkConfig commons.CheckConfig
-	checkConfig.Init(s, c)
+	checkConfig.Init(c)
 	var checks []commons.Check
 	buckets := GetListS3(s)
 	bucketsNotInRegion := GetListS3NotInRegion(s, s.Region)
@@ -53,9 +54,10 @@ func RunChecks(wa *sync.WaitGroup, s aws.Config, c *commons.Config, queue chan [
 	S3toPublicBlockAccess := GetS3ToPublicBlockAccess(s, OnlyBucketInRegion)
 	S3ToVersioning := GetS3ToVersioning(s, OnlyBucketInRegion)
 	S3ToObjectLock := GetS3ToObjectLock(s, OnlyBucketInRegion)
+	S3ToReplicationOtherRegion := GetS3ToReplicationOtherRegion(s, OnlyBucketInRegion)
 
 	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_001", checkIfEncryptionEnabled)(checkConfig, S3ToEncryption, "AWS_S3_001")
-	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_002", CheckIfBucketInOneZone)(checkConfig, couple, "AWS_S3_002")
+	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_002", CheckIfBucketNoReplicationOtherRegion)(checkConfig, S3ToReplicationOtherRegion, "AWS_S3_002")
 	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_003", CheckIfBucketObjectVersioningEnabled)(checkConfig, S3ToVersioning, "AWS_S3_003")
 	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_004", CheckIfObjectLockConfigurationEnabled)(checkConfig, S3ToObjectLock, "AWS_S3_004")
 	go commons.CheckTest(checkConfig.Wg, c, "AWS_S3_005", CheckIfS3PublicAccessBlockEnabled)(checkConfig, S3toPublicBlockAccess, "AWS_S3_005")